Create load_tests function and move old statements to there
