+++ unitt/src/display.rs
@@ -1,8 +1,6 @@
-use std::{fs, path::PathBuf};
+use std;
 
-use glob::glob;
-
-use crate::models::config::Config;
+use crate::collector::LoadedTest;
 use crate::models::statistics::Statistics;
 use crate::models::test;
 
@@ -43,15 +41,20 @@ pub fn print_test_result(test: &test::Test) {
 
 }
 
-pub fn display_tests(config: &Config) -> Summary {
+
+pub fn display_tests<'a, I>(loaded_tests: I) -> Summary
+where
+    I: IntoIterator<Item = LoadedTest>,
+{
     let mut total_stats = Statistics { passed: 0, failed: 0, skipped: 0 };
     let mut file_count = 0u64;
-    for ModuleStreamItem {filename, module} in all_modules_of(config) {
-
+    for LoadedTest {filename, module} in loaded_tests {
         println!("\n===== {} =====\n", filename);
+        
         for test in &module.standalone {
             print_test_result(test);
         }
+        
         for spec in &module.specs {
             println!("\nSuite: {} \n", spec.description);
             for test in &spec.tests {
@@ -71,34 +74,9 @@ pub fn display_tests(config: &Config) -> Summary {
         total_stats.failed += stats.failed;
         total_stats.skipped += stats.skipped;
         file_count += 1;
+
     }
 
     Summary { status: total_stats, file_count }
-}
 
-struct ModuleStreamItem {
-    pub filename: String,
-    pub module: test::Module,
-}
-
-
-fn all_modules_of(config: &Config) -> impl Iterator<Item = ModuleStreamItem> {
-    let pattern = config.target.clone();
-    let files: Vec<_> = glob(&pattern).unwrap().filter_map(Result::ok).collect();
-    let cache = config.cache.clone();
-
-    files.into_iter().map(move |file| {
-        let module = module_from_path(cache.clone(), &file);
-
-        ModuleStreamItem {
-            filename: file.file_name().unwrap().to_string_lossy().into(),
-            module,
-        }
-    })
-}
-
-fn module_from_path(cache: String, file: &PathBuf) -> test::Module {
-    let result_file: PathBuf = PathBuf::from(&cache).join(format!("{}.json", file.to_string_lossy()));
-    let json: String = fs::read_to_string(&result_file).unwrap_or_default();
-    test::Module::from_json(&json)
 }
