+++ unitt/src/main.rs
@@ -1,5 +1,6 @@
 use clap::Parser;
 mod cli;
+mod collector;
 mod display;
 mod models;
 mod runner;
@@ -21,8 +22,10 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
     runner::reset_cache(config.cache.clone());
     runner::generate_tests(&pattern, &arturo).await;
 
+    let tests = collector::load_tests(&config);
+
     println!("\nFinal Summary:");
-    println!("{}", display::display_tests(&config));
+    println!("{}", display::display_tests(tests));
 
     Ok(())
 }
